@@ -1,14 +1,16 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 32
     defaultConfig {
         applicationId "com.lonelypluto.pdfviewerdemo"
         minSdkVersion 17
-        targetSdkVersion 28
+        targetSdkVersion 32
         versionCode 1
         versionName "1.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+
+        ndk.abiFilters 'armeabi-v7a', 'x86'
+        testInstrumentationRunner 'androidx.test.runner.AndroidJUnitRunner'
     }
     buildTypes {
         release {
@@ -16,15 +18,39 @@ android {
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_11
+        targetCompatibility JavaVersion.VERSION_11
+    }
+//    sourceSets {
+//        main {
+//            manifest.srcFile 'src/main/AndroidManifest.xml'
+//            java.srcDirs 'jniLibs', 'src'
+//            res.srcDirs 'res'
+//        }
+//    }
+//    externalNativeBuild {
+//        cmake {
+//            version '3.18.1'
+//        }
+//        ndkBuild.path 'jni/Android.mk'
+//    }
+//    ndkVersion '25.0.8355429'
 }
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:appcompat-v7:28.0.0'
-    implementation 'com.android.support.constraint:constraint-layout:1.1.3'
+    implementation 'androidx.core:core:1.7.0'
+    implementation("androidx.appcompat:appcompat:1.4.1")
+    implementation 'androidx.customview:customview:1.1.0'
+    implementation 'com.google.android.material:material:1.6.0'
+//    implementation 'com.android.support:appcompat-v7:28.0.0'
+    implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
     testImplementation 'junit:junit:4.12'
-    androidTestImplementation 'com.android.support.test:runner:1.0.2'
-    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
+        implementation("com.github.librepdf:openpdf:1.3.23")
+    implementation("ro.andob.androidawt:androidawt:1.0.4")
+//    androidTestImplementation 'com.android.support.test:runner:1.0.2'
+//    androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
 
-    implementation 'com.lonelypluto:pdf-viewer:1.0.7'
+//    implementation 'com.lonelypluto:pdf-viewer:1.0.7'
 }
