@@ -13,8 +13,10 @@ import android.os.AsyncTask;
 import android.os.Build;
 import android.os.Bundle;
 import android.os.Environment;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.core.content.ContextCompat;
+
 import android.text.Editable;
 import android.text.TextWatcher;
 import android.util.DisplayMetrics;
@@ -171,8 +173,8 @@ public class AllPDFActivity extends AppCompatActivity {
         // 判断如果core为空，提示不能打开文件
         if (muPDFCore == null) {
             AlertDialog alert = mAlertBuilder.create();
-            alert.setTitle(com.lonelypluto.pdflibrary.R.string.cannot_open_document);
-            alert.setButton(AlertDialog.BUTTON_POSITIVE, getString(com.lonelypluto.pdflibrary.R.string.dismiss),
+            alert.setTitle(com.lonelypluto.pdfviewerdemo.R.string.cannot_open_document);
+            alert.setButton(AlertDialog.BUTTON_POSITIVE, getString(com.lonelypluto.pdfviewerdemo.R.string.dismiss),
                     new DialogInterface.OnClickListener() {
                         public void onClick(DialogInterface dialog, int which) {
                             finish();
@@ -213,7 +215,7 @@ public class AllPDFActivity extends AppCompatActivity {
         btn_linkhighlightcolor.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                setLinkHighlightColor(ContextCompat.getColor(AllPDFActivity.this, com.lonelypluto.pdflibrary.R.color.link_bg));
+                setLinkHighlightColor(ContextCompat.getColor(AllPDFActivity.this, com.lonelypluto.pdfviewerdemo.R.color.link_bg));
             }
         });
 
@@ -222,7 +224,7 @@ public class AllPDFActivity extends AppCompatActivity {
         btn_searchtextcolor.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                setSearchTextColor(ContextCompat.getColor(AllPDFActivity.this, com.lonelypluto.pdflibrary.R.color.search_bg));
+                setSearchTextColor(ContextCompat.getColor(AllPDFActivity.this, com.lonelypluto.pdfviewerdemo.R.color.search_bg));
             }
         });
 
@@ -231,7 +233,7 @@ public class AllPDFActivity extends AppCompatActivity {
         btn_paintcolor.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
-                int c = ContextCompat.getColor(AllPDFActivity.this, com.lonelypluto.pdflibrary.R.color.rv_item_line_bg);
+                int c = ContextCompat.getColor(AllPDFActivity.this, com.lonelypluto.pdfviewerdemo.R.color.rv_item_line_bg);
                 Log.e(TAG, "color = " + c);
                 setColor(c);
                 setInkColor(0xFF0000FF);
@@ -284,7 +286,7 @@ public class AllPDFActivity extends AppCompatActivity {
                 float density = metric.density;
                 savePdf.setDensity(density);
 
-                Bitmap bitmap = getBitmap(AllPDFActivity.this, com.lonelypluto.pdflibrary.R.mipmap.ic_launcher);
+                Bitmap bitmap = getBitmap(AllPDFActivity.this, com.lonelypluto.pdfviewerdemo.R.mipmap.ic_launcher);
                 savePdf.setBitmap(bitmap);
 
 //                Bitmap bitmap = Bitmap.createBitmap(vdhDeepLayout.getWidth(), vdhDeepLayout.getHeight(),
@@ -445,7 +447,7 @@ public class AllPDFActivity extends AppCompatActivity {
 
         muPDFReaderView.setAdapter(new MuPDFPageAdapter(this, muPDFCore));
         // 设置view的背景色
-        muPDFReaderView.setBackgroundColor(ContextCompat.getColor(this, com.lonelypluto.pdflibrary.R.color.muPDFReaderView_bg));
+        muPDFReaderView.setBackgroundColor(ContextCompat.getColor(this, com.lonelypluto.pdfviewerdemo.R.color.muPDFReaderView_bg));
 
         mSearchTask = new SearchTask(this, muPDFCore) {
             @Override
@@ -725,8 +727,8 @@ public class AllPDFActivity extends AppCompatActivity {
         mTopBarSwitcher.setDisplayedChild(mTopBarMode.ordinal());
         mAcceptMode = AcceptMode.CopyText;
         muPDFReaderView.setMode(MuPDFReaderView.Mode.Selecting);
-        mAnnotTypeText.setText(getString(com.lonelypluto.pdflibrary.R.string.copy_text));
-        showInfo(getString(com.lonelypluto.pdflibrary.R.string.select_text));
+        mAnnotTypeText.setText(getString(com.lonelypluto.pdfviewerdemo.R.string.copy_text));
+        showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.select_text));
     }
 
     /**
@@ -783,8 +785,8 @@ public class AllPDFActivity extends AppCompatActivity {
         mTopBarSwitcher.setDisplayedChild(mTopBarMode.ordinal());
         mAcceptMode = AcceptMode.Highlight;
         muPDFReaderView.setMode(MuPDFReaderView.Mode.Selecting);
-        mAnnotTypeText.setText(com.lonelypluto.pdflibrary.R.string.pdf_tools_highlight);
-        showInfo(getString(com.lonelypluto.pdflibrary.R.string.select_text));
+        mAnnotTypeText.setText(com.lonelypluto.pdfviewerdemo.R.string.pdf_tools_highlight);
+        showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.select_text));
     }
 
     /**
@@ -796,8 +798,8 @@ public class AllPDFActivity extends AppCompatActivity {
         mTopBarSwitcher.setDisplayedChild(mTopBarMode.ordinal());
         mAcceptMode = AcceptMode.Underline;
         muPDFReaderView.setMode(MuPDFReaderView.Mode.Selecting);
-        mAnnotTypeText.setText(com.lonelypluto.pdflibrary.R.string.pdf_tools_underline);
-        showInfo(getString(com.lonelypluto.pdflibrary.R.string.select_text));
+        mAnnotTypeText.setText(com.lonelypluto.pdfviewerdemo.R.string.pdf_tools_underline);
+        showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.select_text));
     }
 
     /**
@@ -809,8 +811,8 @@ public class AllPDFActivity extends AppCompatActivity {
         mTopBarSwitcher.setDisplayedChild(mTopBarMode.ordinal());
         mAcceptMode = AcceptMode.StrikeOut;
         muPDFReaderView.setMode(MuPDFReaderView.Mode.Selecting);
-        mAnnotTypeText.setText(com.lonelypluto.pdflibrary.R.string.pdf_tools_strike_out);
-        showInfo(getString(com.lonelypluto.pdflibrary.R.string.select_text));
+        mAnnotTypeText.setText(com.lonelypluto.pdfviewerdemo.R.string.pdf_tools_strike_out);
+        showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.select_text));
     }
 
     /**
@@ -822,8 +824,8 @@ public class AllPDFActivity extends AppCompatActivity {
         mTopBarSwitcher.setDisplayedChild(mTopBarMode.ordinal());
         mAcceptMode = AcceptMode.Ink;
         muPDFReaderView.setMode(MuPDFReaderView.Mode.Drawing);
-        mAnnotTypeText.setText(com.lonelypluto.pdflibrary.R.string.pdf_tools_ink);
-        showInfo(getString(com.lonelypluto.pdflibrary.R.string.pdf_tools_draw_annotation));
+        mAnnotTypeText.setText(com.lonelypluto.pdfviewerdemo.R.string.pdf_tools_ink);
+        showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.pdf_tools_draw_annotation));
     }
 
     /**
@@ -884,7 +886,7 @@ public class AllPDFActivity extends AppCompatActivity {
                 if (pageView != null)
                     success = pageView.copySelection();
                 mTopBarMode = TopBarMode.Main;
-                showInfo(success ? getString(com.lonelypluto.pdflibrary.R.string.copied_to_clipboard) : getString(com.lonelypluto.pdflibrary.R.string.no_text_selected));
+                showInfo(success ? getString(com.lonelypluto.pdfviewerdemo.R.string.copied_to_clipboard) : getString(com.lonelypluto.pdfviewerdemo.R.string.no_text_selected));
                 break;
             case Highlight:
                 // 高亮
@@ -893,7 +895,7 @@ public class AllPDFActivity extends AppCompatActivity {
                 }
                 mTopBarMode = TopBarMode.Annot;
                 if (!success) {
-                    showInfo(getString(com.lonelypluto.pdflibrary.R.string.no_text_selected));
+                    showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.no_text_selected));
                 }
                 break;
             case Underline:
@@ -901,7 +903,7 @@ public class AllPDFActivity extends AppCompatActivity {
                     success = pageView.markupSelection(Annotation.Type.UNDERLINE);
                 mTopBarMode = TopBarMode.Annot;
                 if (!success)
-                    showInfo(getString(com.lonelypluto.pdflibrary.R.string.no_text_selected));
+                    showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.no_text_selected));
                 break;
 
             case StrikeOut:
@@ -909,7 +911,7 @@ public class AllPDFActivity extends AppCompatActivity {
                     success = pageView.markupSelection(Annotation.Type.STRIKEOUT);
                 mTopBarMode = TopBarMode.Annot;
                 if (!success)
-                    showInfo(getString(com.lonelypluto.pdflibrary.R.string.no_text_selected));
+                    showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.no_text_selected));
                 break;
 
             case Ink:
@@ -917,7 +919,7 @@ public class AllPDFActivity extends AppCompatActivity {
                     success = pageView.saveDraw();
                 mTopBarMode = TopBarMode.Annot;
                 if (!success)
-                    showInfo(getString(com.lonelypluto.pdflibrary.R.string.nothing_to_save));
+                    showInfo(getString(com.lonelypluto.pdfviewerdemo.R.string.nothing_to_save));
                 break;
         }
         mTopBarSwitcher.setDisplayedChild(mTopBarMode.ordinal());
@@ -997,10 +999,10 @@ public class AllPDFActivity extends AppCompatActivity {
     private void showInfo(String message) {
 
         LayoutInflater inflater = getLayoutInflater();
-        View toastLayout = inflater.inflate(com.lonelypluto.pdflibrary.R.layout.toast,
-                (ViewGroup) findViewById(com.lonelypluto.pdflibrary.R.id.toast_root_view));
+        View toastLayout = inflater.inflate(com.lonelypluto.pdfviewerdemo.R.layout.toast,
+                (ViewGroup) findViewById(com.lonelypluto.pdfviewerdemo.R.id.toast_root_view));
 
-        TextView header = (TextView) toastLayout.findViewById(com.lonelypluto.pdflibrary.R.id.toast_message);
+        TextView header = (TextView) toastLayout.findViewById(com.lonelypluto.pdfviewerdemo.R.id.toast_message);
         header.setText(message);
 
         Toast toast = new Toast(getApplicationContext());
@@ -1097,19 +1099,19 @@ public class AllPDFActivity extends AppCompatActivity {
                 }
                 switch (result.buttonGroupType) {
                     case OkCancel:
-                        mAlertDialog.setButton(AlertDialog.BUTTON2, getString(com.lonelypluto.pdflibrary.R.string.cancel), listener);
+                        mAlertDialog.setButton(AlertDialog.BUTTON2, getString(com.lonelypluto.pdfviewerdemo.R.string.cancel), listener);
                         pressed[1] = MuPDFAlert.ButtonPressed.Cancel;
                     case Ok:
-                        mAlertDialog.setButton(AlertDialog.BUTTON1, getString(com.lonelypluto.pdflibrary.R.string.okay), listener);
+                        mAlertDialog.setButton(AlertDialog.BUTTON1, getString(com.lonelypluto.pdfviewerdemo.R.string.okay), listener);
                         pressed[0] = MuPDFAlert.ButtonPressed.Ok;
                         break;
                     case YesNoCancel:
-                        mAlertDialog.setButton(AlertDialog.BUTTON3, getString(com.lonelypluto.pdflibrary.R.string.cancel), listener);
+                        mAlertDialog.setButton(AlertDialog.BUTTON3, getString(com.lonelypluto.pdfviewerdemo.R.string.cancel), listener);
                         pressed[2] = MuPDFAlert.ButtonPressed.Cancel;
                     case YesNo:
-                        mAlertDialog.setButton(AlertDialog.BUTTON1, getString(com.lonelypluto.pdflibrary.R.string.yes), listener);
+                        mAlertDialog.setButton(AlertDialog.BUTTON1, getString(com.lonelypluto.pdfviewerdemo.R.string.yes), listener);
                         pressed[0] = MuPDFAlert.ButtonPressed.Yes;
-                        mAlertDialog.setButton(AlertDialog.BUTTON2, getString(com.lonelypluto.pdflibrary.R.string.no), listener);
+                        mAlertDialog.setButton(AlertDialog.BUTTON2, getString(com.lonelypluto.pdfviewerdemo.R.string.no), listener);
                         pressed[1] = MuPDFAlert.ButtonPressed.No;
                         break;
                 }
@@ -1210,10 +1212,10 @@ public class AllPDFActivity extends AppCompatActivity {
                 }
             };
             AlertDialog alert = mAlertBuilder.create();
-            alert.setTitle(com.lonelypluto.pdflibrary.R.string.dialog_title);
-            alert.setMessage(getString(com.lonelypluto.pdflibrary.R.string.document_has_changes_save_them));
-            alert.setButton(AlertDialog.BUTTON_POSITIVE, getString(com.lonelypluto.pdflibrary.R.string.yes), listener);
-            alert.setButton(AlertDialog.BUTTON_NEGATIVE, getString(com.lonelypluto.pdflibrary.R.string.no), listener);
+            alert.setTitle(com.lonelypluto.pdfviewerdemo.R.string.dialog_title);
+            alert.setMessage(getString(com.lonelypluto.pdfviewerdemo.R.string.document_has_changes_save_them));
+            alert.setButton(AlertDialog.BUTTON_POSITIVE, getString(com.lonelypluto.pdfviewerdemo.R.string.yes), listener);
+            alert.setButton(AlertDialog.BUTTON_NEGATIVE, getString(com.lonelypluto.pdfviewerdemo.R.string.no), listener);
             alert.show();
         } else {
             finish();
