@@ -1,11 +1,12 @@
 package com.lonelypluto.pdfviewerdemo.adapter;
 
 import android.content.Context;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import com.artifex.mupdfdemo.OutlineItem;
 import com.lonelypluto.pdfviewerdemo.R;
